requirements.txt
Django==4.1
djangorestframework==3.13.1
django-filter==22.1
psycopg2-binary==2.9.3

# tests
factory-boy==3.2.1
pytest==7.1.2
pytest-django==4.5.2
prospector==1.7.7
